Adding a '-server-mode' mode to snapshot (both mysqlctl and vtctl). In that mode, we serve the raw mysql data files while mysql is stopped, compressing them on the fly.
LGTM Mike.

+++ go/vt/mysqlctl/fileutil.go
@@ -17,6 +17,7 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"strings"
 )
 
 // Use this to simulate failures in tests
@@ -30,13 +31,15 @@ func init() {
 	simulateFailures = statErr == nil
 }
 
-// compressFile compresses a single file with gzip, leaving the src
-// file intact. The path of the returned SnapshotFile will be relative
+// newSnapshotFile behavior depends on the compress flag:
+// - if compress is true , it compresses a single file with gzip, and
+// computes the md5 on the compressed version.
+// - if compress is false, just symlinks and computes the md5 on the file
+// The source file is always left intact.
+// The path of the returned SnapshotFile will be relative
 // to root. Also computes the md5 hash on the fly.
-func compressFile(srcPath, dstPath, root string) (*SnapshotFile, error) {
-	// FIXME(msolomon) not sure how well Go will schedule cpu intensive tasks
-	// might be better if this forked off workers.
-	relog.Info("compressFile: starting to compress %v into %v", srcPath, dstPath)
+func newSnapshotFile(srcPath, dstPath, root string, compress bool) (*SnapshotFile, error) {
+	relog.Info("newSnapshotFile: starting to compress %v into %v", srcPath, dstPath)
 	srcFile, err := os.OpenFile(srcPath, os.O_RDONLY, 0)
 	if err != nil {
 		return nil, err
@@ -45,43 +48,64 @@ func compressFile(srcPath, dstPath, root string) (*SnapshotFile, error) {
 
 	src := bufio.NewReaderSize(srcFile, 2*1024*1024)
 
-	dir, filePrefix := path.Split(dstPath)
+	var hash string
+	if compress {
+		relog.Info("newSnapshotFile: starting to compress %v into %v", srcPath, dstPath)
+		dir, filePrefix := path.Split(dstPath)
 
-	dstFile, err := ioutil.TempFile(dir, filePrefix)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		// try to close and delete the file.
-		// in the success case, the file will already be closed
-		// and renamed, so all of this would fail anyway, no biggie
+		dstFile, err := ioutil.TempFile(dir, filePrefix)
+		if err != nil {
+			return nil, err
+		}
+		defer func() {
+			// try to close and delete the file.  in the
+			// success case, the file will already be
+			// closed and renamed, so all of this would
+			// fail anyway, no biggie
+			dstFile.Close()
+			os.Remove(dstFile.Name())
+		}()
+
+		dst := bufio.NewWriterSize(dstFile, 2*1024*1024)
+
+		hasher := md5.New()
+		tee := io.MultiWriter(dst, hasher)
+
+		compressor := gzip.NewWriter(tee)
+		defer compressor.Close()
+
+		_, err = io.Copy(compressor, src)
+		if err != nil {
+			return nil, err
+		}
+
+		// close dst manually to flush all buffers to disk
+		compressor.Close()
+		dst.Flush()
 		dstFile.Close()
-		os.Remove(dstFile.Name())
-	}()
+		hash = hex.EncodeToString(hasher.Sum(nil))
 
-	dst := bufio.NewWriterSize(dstFile, 2*1024*1024)
+		// atomically move completed compressed file
+		err = os.Rename(dstFile.Name(), dstPath)
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		relog.Info("newSnapshotFile: starting to hash and symlinking %v to %v", srcPath, dstPath)
 
-	hasher := md5.New()
-	tee := io.MultiWriter(dst, hasher)
+		// get the md5
+		hasher := md5.New()
+		_, err = io.Copy(hasher, src)
+		if err != nil {
+			return nil, err
+		}
+		hash = hex.EncodeToString(hasher.Sum(nil))
 
-	compressor := gzip.NewWriter(tee)
-	defer compressor.Close()
-
-	_, err = io.Copy(compressor, src)
-	if err != nil {
-		return nil, err
-	}
-
-	// close dst manually to flush all buffers to disk
-	compressor.Close()
-	dst.Flush()
-	dstFile.Close()
-	hash := hex.EncodeToString(hasher.Sum(nil))
-
-	// atomically move completed compressed file
-	err = os.Rename(dstFile.Name(), dstPath)
-	if err != nil {
-		return nil, err
+		// do the symlink
+		err = os.Symlink(srcPath, dstPath)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	relog.Info("clone data ready %v:%v", dstPath, hash)
@@ -92,9 +116,13 @@ func compressFile(srcPath, dstPath, root string) (*SnapshotFile, error) {
 	return &SnapshotFile{relativeDst, hash}, nil
 }
 
-// compressFile compresses multiple files in parallel. The Paths of
+// newSnapshotFiles processes multiple files in parallel. The Paths of
 // the returned SnapshotFiles will be relative to root.
-func compressFiles(sources, destinations []string, root string, compressConcurrency int) ([]SnapshotFile, error) {
+// - if compress is true, we compress the files and compute the md5 on
+// the compressed version.
+// - if compress is false, we symlink the files, and compute the md5 on
+// the original version.
+func newSnapshotFiles(sources, destinations []string, root string, concurrency int, compress bool) ([]SnapshotFile, error) {
 	if len(sources) != len(destinations) || len(sources) == 0 {
 		panic(fmt.Errorf("bad array lengths: %v %v", len(sources), len(destinations)))
 	}
@@ -107,10 +135,10 @@ func compressFiles(sources, destinations []string, root string, compressConcurre
 
 	snapshotFiles := make([]SnapshotFile, len(sources))
 	resultQueue := make(chan error, len(sources))
-	for i := 0; i < compressConcurrency; i++ {
+	for i := 0; i < concurrency; i++ {
 		go func() {
 			for i := range workQueue {
-				sf, err := compressFile(sources[i], destinations[i], root)
+				sf, err := newSnapshotFile(sources[i], destinations[i], root, compress)
 				if err == nil {
 					snapshotFiles[i] = *sf
 				}
@@ -147,7 +175,7 @@ func compressFiles(sources, destinations []string, root string, compressConcurre
 // a gunzip reader writing to a file.  It will compare the md5
 // checksum after the copy is done.
 func fetchFile(srcUrl, srcHash, dstFilename string) error {
-	relog.Info("fetchFile: starting to fetch %v", dstFilename)
+	relog.Info("fetchFile: starting to fetch %v from %v", dstFilename, srcUrl)
 
 	// create destination directory
 	dir, _ := path.Split(dstFilename)
@@ -186,13 +214,18 @@ func fetchFile(srcUrl, srcHash, dstFilename string) error {
 	// and into the gunziper
 	tee := io.TeeReader(resp.Body, hasher)
 
-	// FIXME(msolomon) only uncompress if url ends with .gz
 	// create the uncompresser
-	decompressor, err := gzip.NewReader(tee)
-	if err != nil {
-		return err
+	var decompressor io.Reader
+	if strings.HasSuffix(srcUrl, ".gz") {
+		gz, err := gzip.NewReader(tee)
+		if err != nil {
+			return err
+		}
+		defer gz.Close()
+		decompressor = gz
+	} else {
+		decompressor = tee
 	}
-	defer decompressor.Close()
 
 	// see if we need to introduce failures
 	if simulateFailures {
